Update SonarQube
Former-commit-id: 1a86999f458701024723f63beedad3a436e0cf8e [formerly 83d622b5e0eddda3060cbb6fbb96edf539c83676]
Former-commit-id: a4f13f1d9d5fb92474665d7a813aeac838bbd7f7

+++ backend/RunWithMe/src/main/java/kr/co/rwm/service/ReplyServiceImpl.java
@@ -85,13 +85,18 @@ public class ReplyServiceImpl implements ReplyService {
 	@Override
 	@Transactional
 	public Long delete(int reply_id) {
-		int boardId = replyRepository.findByReplyId(reply_id).get().getBoardId();
-		Optional<Board> board = boardRepository.findByBoardId(boardId);
-		if(board.isPresent()) {
-			board.get().setReplyCount(board.get().getReplyCount()-1);
-			Long ret = replyRepository.deleteByReplyId(reply_id);
-			return ret;
-			
+		Optional<Reply> comment = replyRepository.findByReplyId(reply_id);
+		if(comment.isPresent()) {
+			int boardId = comment.get().getBoardId();
+			Optional<Board> board = boardRepository.findByBoardId(boardId);
+			if(board.isPresent()) {
+				board.get().setReplyCount(board.get().getReplyCount()-1);
+				Long ret = replyRepository.deleteByReplyId(reply_id);
+				return ret;
+				
+			}else {
+				return null;
+			}
 		}else {
 			return null;
 		}
